requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
aiofiles>=23.2.0
httpx>=0.25.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
